Sport-Front/src/app/app.component.ts
import { Component, OnInit } from '@angular/core';

/**
 * The app component. This component is the base of the BookStore
 */
@Component({
    selector: 'app-root',
    templateUrl: './app.component.html',
    styleUrls: ['./app.component.css']
})
export class AppComponent implements OnInit {

    /**
     * The title that appears on the NavBar and the web browser
     */
    title: String;

    /**
     * Assigns a title to the web page
     */
    ngOnInit(): void {
        this.title = "BookStore";
    }

    /**
     * @ignore
     */
    constructor() { }

}





